UPDATE-Task-35 - Atualiza o end-point de resumo para retornar um responseEntity e ajusta o gastoResumoResponseDTO

+++ src/main/java/br/com/walletfy/dto/GastoResumoResponseDTO.java
@@ -16,7 +16,7 @@ public class GastoResumoResponseDTO {
 	BigDecimal receita;
 	BigDecimal despesas;
 	BigDecimal saldo;
-	Integer pendentes;
+	Long pendentes;
 	BigDecimal variacaoReceita;
 	BigDecimal variacaoDespesas;
 	BigDecimal variacaoSaldo;
